+++ src/components/SwitchActiveAffiliation/index.js
@@ -1,0 +1,1 @@
+export { SwitchActiveAffiliation } from './SwitchActiveAffiliation';