+++ src/constants.js
@@ -1,0 +1,9 @@
+export const CONTENT_TYPE_HEADER = 'Content-Type';
+export const OKAPI_TENANT_HEADER = 'X-Okapi-Tenant';
+export const OKAPI_TOKEN_HEADER = 'X-Okapi-Token';
+
+export const CONFIGURATIONS_API = 'configurations';
+export const CONFIGURATIONS_ENTRIES_API = `${CONFIGURATIONS_API}/entries`;
+export const CONSORTIA_API = 'consortia';
+export const CONSORTIA_TENANTS_API = 'tenants';
+export const CONSORTIA_USER_TENANTS_API = 'user-tenants';