+++ src/handlers/handleChangeAffilication.js
@@ -1,0 +1,3 @@
+import { SwitchActiveAffiliation } from '../components';
+
+export const handleChangeAffilication = () => SwitchActiveAffiliation;